asr-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.2
openai-whisper==20250625
librosa==0.10.1
structlog==23.2.0
python-dotenv==1.0.0
